plugin Weather Underground ========================== - Add Rain in 1 hour, in the day. - Add the Visibility. - correct comments in UV - Delete keywords in Fahrenheit. Only Keywords in Celcius are stored in the DB.

+++ sources/plugins/WUnderGround/Keywords/UV.cpp
@@ -26,7 +26,7 @@ CUV::~CUV()
 
 void CUV::GetValue(double temp)
 {
-   m_uv.set(temp);
+   m_uv.set( (int) temp);
 }
 
 void CUV::DeclareKeywords (boost::shared_ptr<yApi::IYPluginApi> context ) const
